data: install tmpfiles.d/image-builder.conf to auto-clean cache
Our osbuild/image-builder cache is located in /var/cache/image-builder
and it can grow quite big. To avoid it growing out of bounds too much
add a tmpfile.d snippet to ensure that we clean content older than
14d. The 14d is a bit arbitrary, happy to change to longer or shorter
cache timeouts (side-note: we have a cache limit in osbuild and an LRU
there too but sources are currently not part of the cache limits in
osbuild).